+++ scripts/snapshots/es6Files.txt
@@ -53,7 +53,7 @@
   "es6/util/cursor/getCursorPoints.js",
   "es6/state/tooltipSlice.js",
   "es6/state/store.js",
-  "es6/state/selectors.js",
+  "es6/state/selectors",
   "es6/state/referenceElementsSlice.js",
   "es6/state/reduxDevtoolsJsonStringifyReplacer.js",
   "es6/state/optionsSlice.js",
@@ -61,15 +61,17 @@
   "es6/state/layoutSlice.js",
   "es6/state/hooks.js",
   "es6/state/graphicalItemsSlice.js",
-  "es6/state/dataSelectors.js",
   "es6/state/chartDataSlice.js",
   "es6/state/axisSlice.js",
-  "es6/state/axisSelectors.js",
   "es6/state/axisMapSlice.js",
   "es6/state/SetTooltipEntrySettings.js",
   "es6/state/SetCartesianGraphicalItem.js",
   "es6/state/ReportBar.js",
   "es6/state/RechartsStoreProvider.js",
+  "es6/state/selectors/selectors.js",
+  "es6/state/selectors/dataSelectors.js",
+  "es6/state/selectors/containerSelectors.js",
+  "es6/state/selectors/axisSelectors.js",
   "es6/shape/Trapezoid.js",
   "es6/shape/Symbols.js",
   "es6/shape/Sector.js",
