+++ scripts/snapshots/libFiles.txt
@@ -53,7 +53,7 @@
   "lib/util/cursor/getCursorPoints.js",
   "lib/state/tooltipSlice.js",
   "lib/state/store.js",
-  "lib/state/selectors.js",
+  "lib/state/selectors",
   "lib/state/referenceElementsSlice.js",
   "lib/state/reduxDevtoolsJsonStringifyReplacer.js",
   "lib/state/optionsSlice.js",
@@ -61,15 +61,17 @@
   "lib/state/layoutSlice.js",
   "lib/state/hooks.js",
   "lib/state/graphicalItemsSlice.js",
-  "lib/state/dataSelectors.js",
   "lib/state/chartDataSlice.js",
   "lib/state/axisSlice.js",
-  "lib/state/axisSelectors.js",
   "lib/state/axisMapSlice.js",
   "lib/state/SetTooltipEntrySettings.js",
   "lib/state/SetCartesianGraphicalItem.js",
   "lib/state/ReportBar.js",
   "lib/state/RechartsStoreProvider.js",
+  "lib/state/selectors/selectors.js",
+  "lib/state/selectors/dataSelectors.js",
+  "lib/state/selectors/containerSelectors.js",
+  "lib/state/selectors/axisSelectors.js",
   "lib/shape/Trapezoid.js",
   "lib/shape/Symbols.js",
   "lib/shape/Sector.js",
