+++ scripts/snapshots/typesFiles.txt
@@ -53,7 +53,7 @@
   "types/util/cursor/getCursorPoints.d.ts",
   "types/state/tooltipSlice.d.ts",
   "types/state/store.d.ts",
-  "types/state/selectors.d.ts",
+  "types/state/selectors",
   "types/state/referenceElementsSlice.d.ts",
   "types/state/reduxDevtoolsJsonStringifyReplacer.d.ts",
   "types/state/optionsSlice.d.ts",
@@ -61,15 +61,17 @@
   "types/state/layoutSlice.d.ts",
   "types/state/hooks.d.ts",
   "types/state/graphicalItemsSlice.d.ts",
-  "types/state/dataSelectors.d.ts",
   "types/state/chartDataSlice.d.ts",
   "types/state/axisSlice.d.ts",
-  "types/state/axisSelectors.d.ts",
   "types/state/axisMapSlice.d.ts",
   "types/state/SetTooltipEntrySettings.d.ts",
   "types/state/SetCartesianGraphicalItem.d.ts",
   "types/state/ReportBar.d.ts",
   "types/state/RechartsStoreProvider.d.ts",
+  "types/state/selectors/selectors.d.ts",
+  "types/state/selectors/dataSelectors.d.ts",
+  "types/state/selectors/containerSelectors.d.ts",
+  "types/state/selectors/axisSelectors.d.ts",
   "types/shape/Trapezoid.d.ts",
   "types/shape/Symbols.d.ts",
   "types/shape/Sector.d.ts",
